Correct datatype of method and change rounding
The method calcCoverageWithoutN(OverallQualityAssessmentMerged has got the wrong type long
instead of the correct type double. This is changed here.

Also the output use now a formated output with always two digits after the dot.

+++ grails-app/services/de/dkfz/tbi/otp/dataprocessing/OverallQualityAssessmentMergedService.groovy
@@ -4,7 +4,7 @@ import de.dkfz.tbi.otp.ngsdata.ReferenceGenome
 
 class OverallQualityAssessmentMergedService {
 
-    Long calcCoverageWithoutN(OverallQualityAssessmentMerged overallQualityAssessmentMerged, ReferenceGenome referenceGenome) {
+    Double calcCoverageWithoutN(OverallQualityAssessmentMerged overallQualityAssessmentMerged, ReferenceGenome referenceGenome) {
         long qcBasesMapped = overallQualityAssessmentMerged.qcBasesMapped
         long referenceGenomeLengthWithoutN = referenceGenome.lengthWithoutN
         double coverageWithoutN = qcBasesMapped / referenceGenomeLengthWithoutN
